feat: 장애인 승하차 체크 기능 개발
